refactor(runt-mcp): deduplicate session validation and execution result building
Extract `require_handle!` macro and shared `build_execution_result` to
eliminate repeated boilerplate across all MCP tool files. Net -304 lines
with no behavior changes.
